New refinement method
